+++ lib/active_model/serializer/adapter/json.rb
@@ -1,3 +1,5 @@
+require 'active_model/serializer/adapter/json/fragment_cache'
+
 module ActiveModel
   class Serializer
     class Adapter
@@ -6,30 +8,44 @@ module ActiveModel
           if serializer.respond_to?(:each)
             @result = serializer.map{|s| self.class.new(s).serializable_hash }
           else
-            @result = cached_object do
-              @hash = serializer.attributes(options)
-              serializer.each_association do |name, association, opts|
-                if association.respond_to?(:each)
-                  array_serializer = association
-                  @hash[name] = array_serializer.map { |item| item.attributes(opts) }
-                else
-                  if association
-                    @hash[name] = association.attributes(options)
-                  else
-                    @hash[name] = nil
+            @hash = {}
+
+            @core = cache_check(serializer) do
+              serializer.attributes(options)
+            end
+
+            serializer.each_association do |name, association, opts|
+              if association.respond_to?(:each)
+                array_serializer = association
+                @hash[name] = array_serializer.map do |item|
+                  cache_check(item) do
+                    item.attributes(opts)
                   end
                 end
+              else
+                if association
+                  @hash[name] = cache_check(association) do
+                    association.attributes(options)
+                  end
+                elsif opts[:virtual_value]
+                  @hash[name] = opts[:virtual_value]
+                else
+                  @hash[name] = nil
+                end
               end
-              @hash
             end
+            @result = @core.merge @hash
           end
 
           if root = options.fetch(:root, serializer.json_key)
             @result = { root => @result }
           end
-
           @result
         end
+      end
+
+      def fragment_cache(cached_hash, non_cached_hash)
+        Json::FragmentCache.new().fragment_cache(cached_hash, non_cached_hash)
       end
     end
   end