+++ test/adapter/json_api/has_one_test.rb
@@ -41,6 +41,7 @@ module ActiveModel
             expected = [
               {
                 id: "43",
+                rating: nil,
                 type: "bios",
                 content:"AMS Contributor",
                 links: {