+++ test/adapter/json_test.rb
@@ -5,6 +5,7 @@ module ActiveModel
     class Adapter
       class JsonTest < Minitest::Test
         def setup
+          ActionController::Base.cache_store.clear
           @author = Author.new(id: 1, name: 'Steve K.')
           @post = Post.new(title: 'New Post', body: 'Body')
           @first_comment = Comment.new(id: 1, body: 'ZOMG A COMMENT')