+++ CHANGELOG.md
@@ -4,3 +4,5 @@
   * adds method to override association [adcb99e, @kurko]
   * adds `has_one` attribute for backwards compatibility [@ggordon]
   * updates JSON API support to RC3 [@mateomurphy]
+  * adds fragment cache support [@joaomdmoura]
+  * adds cache support to attributes and associations [@joaomdmoura]